Move test files into `test/unit-tests`
I think these tests got misfiled. All the other jest tests are under `test/unit-tests`.
