+++ test/unit-tests/components/views/dialogs/security/SetupEncryptionDialog-test.tsx
@@ -11,10 +11,10 @@ import { render, screen } from "jest-matrix-react";
 import { type Mocked } from "jest-mock";
 import { type CryptoApi } from "matrix-js-sdk/src/crypto-api";
 
-import SetupEncryptionDialog from "../../../../../src/components/views/dialogs/security/SetupEncryptionDialog";
-import { getMockClientWithEventEmitter } from "../../../../test-utils";
-import { Phase, SetupEncryptionStore } from "../../../../../src/stores/SetupEncryptionStore";
-import Modal from "../../../../../src/Modal";
+import SetupEncryptionDialog from "../../../../../../src/components/views/dialogs/security/SetupEncryptionDialog";
+import { getMockClientWithEventEmitter } from "../../../../../test-utils";
+import { Phase, SetupEncryptionStore } from "../../../../../../src/stores/SetupEncryptionStore";
+import Modal from "../../../../../../src/Modal";
 
 describe("SetupEncryptionDialog", () => {
     afterEach(() => {
